+++ packages/lib/src/actionMiddlewares/actionSerialization/arraySerializer.ts
@@ -5,14 +5,14 @@ import { ActionCallArgumentSerializer, cannotSerialize } from "./core"
 export const arraySerializer: ActionCallArgumentSerializer<any[] | IObservableArray<any>, any[]> = {
   id: "mobx-keystone/array",
 
-  serialize(value, _, serialize) {
+  serialize(value, serialize) {
     if (!isArray(value)) return cannotSerialize
 
     // this will also transform observable arrays into non observable ones
     return value.map(serialize)
   },
 
-  deserialize(arr, _, deserialize) {
+  deserialize(arr, deserialize) {
     return arr.map(deserialize)
   },
 }