+++ packages/lib/src/actionMiddlewares/actionSerialization/plainObjectSerializer.ts
@@ -5,14 +5,14 @@ import { ActionCallArgumentSerializer, cannotSerialize } from "./core"
 export const plainObjectSerializer: ActionCallArgumentSerializer<object, object> = {
   id: "mobx-keystone/plainObject",
 
-  serialize(value, _, serialize) {
+  serialize(value, serialize) {
     if (!isPlainObject(value) && !isObservableObject(value)) return cannotSerialize
 
     // this will make observable objects non observable ones
     return mapObjectFields(value, serialize)
   },
 
-  deserialize(obj, _, serialize) {
+  deserialize(obj, serialize) {
     return mapObjectFields(obj, serialize)
   },
 }