+++ packages/lib/src/propTransform/index.ts
@@ -1,0 +1,2 @@
+export { stringAsDate, timestampAsDate } from "./asDate"
+export { propTransform, PropTransform, PropTransformDecorator } from "./propTransform"