+++ packages/site/doczrc.js
@@ -28,6 +28,7 @@ export default {
     "References",
     "Frozen Data",
     "Runtime Type Checking",
+    "Property Transforms",
     "Redux Compatibility",
     { name: "Examples", menu: ["Todo List", "Client/Server"] },
     "API Reference",