+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 - [BREAKING CHANGE] Added a `$modelId` extra property to models and their snapshots in order to be able to properly validate targets of serialized actions in scenarions with concurrent clients. Also allows the automatic optimization of the serialized version of models in action parameters by substituting them to just their paths + path of ids whenever possible.
 - [BREAKING CHANGE] Default values for properties will now also apply when the initial data is `null`.
 - `serializeActionCall`, `serializeActionCallArgument`, `deserializeActionCall` and `deserializeActionCallArgument` now can take a second parameter with the root node of the model where actions are going to be performed to optimize the serialization of arguments that are already in the store.
+- Added `registerActionCallArgumentSerializer` for serialization of custom action argument types.
 - `ActionContext` now includes `targetPathIds`.
 - Added `pathObjects` to `getRootPath`.
 - Improved a bit the typing for `fromSnapshot` model methods.
+- Added property transforms via `propTransform`.
 - Improved action argument serialization so it supports dates, maps, sets, arrays of models, etc.
 - `arrayAsMap` now supports arbitrary key types.